Menu principal, login y registro

+++ src/Controller/ArticuloController.php
@@ -10,32 +10,42 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
-use Symfony\Component\Security\Http\Attribute\IsGranted;
-use Symfony\Component\HttpFoundation\File\UploadedFile;
 use Symfony\Component\String\Slugger\SluggerInterface;
 
-#[IsGranted('ROLE_ADMIN')]
-#[Route('/admin/articulo')]
+#[Route('/articulo')]
 final class ArticuloController extends AbstractController
 {
     #[Route(name: 'app_articulo_index', methods: ['GET'])]
     public function index(ArticuloRepositorio $articuloRepositorio): Response
     {
+        $usuario = $this->getUser();
+
+        if ($this->isGranted('ROLE_ADMIN')) {
+            $articulos = $articuloRepositorio->findAll();
+        } else {
+            $articulos = $articuloRepositorio->findBy([
+                'usuario' => $usuario
+            ]);
+        }
+
         return $this->render('articulo/index.html.twig', [
-            'articulos' => $articuloRepositorio->findAll(),
+            'articulos' => $articulos,
         ]);
     }
 
     #[Route('/new', name: 'app_articulo_new', methods: ['GET', 'POST'])]
-    public function new(Request $request, EntityManagerInterface $entityManager,
-        SluggerInterface $slugger): Response
-    {
+    public function new(
+        Request $request,
+        EntityManagerInterface $entityManager,
+        SluggerInterface $slugger
+    ): Response {
         $articulo = new Articulo();
         $form = $this->createForm(ArticuloType::class, $articulo);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            
+
+            // Asignar propietario
             $articulo->setUsuario($this->getUser());
 
             $imagenFile = $form->get('imagen')->getData();
@@ -56,7 +66,7 @@ final class ArticuloController extends AbstractController
             $entityManager->persist($articulo);
             $entityManager->flush();
 
-            return $this->redirectToRoute('app_articulo_index', [], Response::HTTP_SEE_OTHER);
+            return $this->redirectToRoute('app_articulo_index');
         }
 
         return $this->render('articulo/new.html.twig', [
@@ -65,24 +75,28 @@ final class ArticuloController extends AbstractController
         ]);
     }
 
-    #[Route('/{id}', name: 'app_articulo_show', methods: ['GET'])]
-    public function show(Articulo $articulo): Response
-    {
-        return $this->render('articulo/show.html.twig', [
-            'articulo' => $articulo,
-        ]);
-    }
-
     #[Route('/{id}/edit', name: 'app_articulo_edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request, Articulo $articulo, EntityManagerInterface $entityManager): Response
-    {
+    public function edit(
+        Request $request,
+        Articulo $articulo,
+        EntityManagerInterface $entityManager
+    ): Response {
+        $usuario = $this->getUser();
+
+        if (
+            !$this->isGranted('ROLE_ADMIN') &&
+            $articulo->getUsuario() !== $usuario
+        ) {
+            throw $this->createAccessDeniedException('No puedes editar este artículo.');
+        }
+
         $form = $this->createForm(ArticuloType::class, $articulo);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $entityManager->flush();
 
-            return $this->redirectToRoute('app_articulo_index', [], Response::HTTP_SEE_OTHER);
+            return $this->redirectToRoute('app_articulo_index');
         }
 
         return $this->render('articulo/edit.html.twig', [
@@ -92,13 +106,25 @@ final class ArticuloController extends AbstractController
     }
 
     #[Route('/{id}', name: 'app_articulo_delete', methods: ['POST'])]
-    public function delete(Request $request, Articulo $articulo, EntityManagerInterface $entityManager): Response
-    {
+    public function delete(
+        Request $request,
+        Articulo $articulo,
+        EntityManagerInterface $entityManager
+    ): Response {
+        $usuario = $this->getUser();
+
+        if (
+            !$this->isGranted('ROLE_ADMIN') &&
+            $articulo->getUsuario() !== $usuario
+        ) {
+            throw $this->createAccessDeniedException('No puedes eliminar este artículo.');
+        }
+
         if ($this->isCsrfTokenValid('delete'.$articulo->getId(), $request->getPayload()->getString('_token'))) {
             $entityManager->remove($articulo);
             $entityManager->flush();
         }
 
-        return $this->redirectToRoute('app_articulo_index', [], Response::HTTP_SEE_OTHER);
+        return $this->redirectToRoute('app_articulo_index');
     }
 }
